rollup and npm scripts
