+++ .babelrc
@@ -1,0 +1,12 @@
+{
+  "presets": [
+    [
+      "env",
+      {
+        "targets": {
+          "node": "4.8.5"
+        }
+      }
+    ]
+  ]
+}
