+++ lib/server.js
@@ -1,6 +1,6 @@
 import Keen from './keen'
 import Request from './request'
-import * as httpHandlers from './utils/http-browser'
+import * as httpHandlers from './utils/http-server'
 
 Request.httpHandlers = httpHandlers
 Keen.Request = Request
