+++ package.json
@@ -2,13 +2,16 @@
   "name": "keen-analysis",
   "version": "1.3.0",
   "description": "Data Analysis SDK for Keen IO",
-  "main": "lib/index.js",
-  "browser": {
-    "./lib/utils/http-server.js": "./lib/utils/http-browser.js"
-  },
+  "main": "dist/server.js",
   "scripts": {
-    "build": "rollup lib/index.js --format umd --name keen-analysis --output dist/keen-analysis.js",
-    "test": "gulp test:unit"
+    "build": "npm run build:server && npm run build:browser && npm run build:minify",
+    "build:browser": "rollup --config rollup.config.js",
+    "build:test": "rollup --config rollup.test.config.js",
+    "build:server": "babel lib --out-dir dist",
+    "build:minify": "uglifyjs dist/browser.js --output dist/browser.min.js --compress --mangle",
+    "test": "npm run test:mocha && npm run test:phantom",
+    "test:mocha": "mocha --reporter nyan --timeout 300000 test/unit/build/index.js",
+    "test:phantom": "mocha-phantomjs --timeout 300000 ./test/unit/index.html -p $(which phantomjs)"
   },
   "repository": {
     "type": "git",
@@ -25,7 +28,10 @@
     "keen-core": "^0.2.0"
   },
   "devDependencies": {
-    "browserify": "^12.0.1",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-plugin-external-helpers": "^6.22.0",
+    "babel-preset-env": "^1.6.1",
     "del": "^2.1.0",
     "gulp": "^3.9.1",
     "gulp-awspublish": "^3.0.1",
@@ -48,10 +54,19 @@
     "karma-safari-launcher": "^0.1.1",
     "karma-sauce-launcher": "^0.3.0",
     "mocha": "^2.3.4",
+    "mocha-phantomjs": "^4.1.0",
     "moment": "^2.10.6",
     "proclaim": "^3.4.0",
     "requirejs": "^2.1.22",
     "rollup": "^0.50.0",
-    "through2": "^2.0.0"
+    "rollup-plugin-alias": "^1.4.0",
+    "rollup-plugin-babel": "^2.7.1",
+    "rollup-plugin-commonjs": "^8.2.6",
+    "rollup-plugin-node-resolve": "^2.1.1",
+    "rollup-plugin-replace": "^2.0.0",
+    "rollup-stream": "^1.24.1",
+    "uglify-js": "^3.1.6",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0"
   }
 }
