+++ test/unit/index.html
@@ -7,7 +7,7 @@
 	<div id="mocha"></div>
   <script src="../vendor/mocha.js"></script>
   <script>mocha.setup('bdd')</script>
-  <script src="./build/browserified-tests.js"></script>
+  <script src="./build/index.js"></script>
   <script type="text/javascript">
   if (window.mochaPhantomJS) {
     mochaPhantomJS.run();
